makegen: fix string concatenation on obj list generation
The variable project_OBJS was first initialized to NULL and then
realloc-ed each time a new object file needed to be added.

When realloc-ing the buffer, if no data is there apparently the buffer
might not be filled with 0's. Causing it to add jibberish at the start
of the buffer, thus corrupting the string.